requirements.txt
Django==2.1.4
Pillow==5.4.0
pytz==2018.7
django-ckeditor==5.6.1
django-wysiwyg==0.8.0